@@ -1,13 +1,13 @@
 import React, { Component } from "react";
 
-export class Libraries extends Component {
+class Home extends Component {
   render() {
     return (
       <div>
-        <h1>Librares</h1>
+        <h1>home</h1>
       </div>
     );
   }
 }
 
-export default Libraries;
+export default Home;
